requirements.txt
requests>=2.25.1
APScheduler>=3.8.1
python-telegram-bot==20.7
psycopg2-binary>=2.9.6